requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
Markdown==3.3.6
mysqlclient==2.1.0
Pillow==9.1.0
pycodestyle==2.8.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
whitenoise==6.0.0
zipp==3.8.0
